Various CODA-M commits squashed together
We should probably provide the entire dir here. [At least that is how
I understand the documentation.]

Initializer for macOS. With this, the HULLO is sent to the app (but it
crashes it ~immediately).

Cosmetics.

Keep the 'document' only in the ViewController. The COWrapper is just
what its name says - a wrapper (between C++ and Swift, via
Objective-C++), so it shouldn't keep a copy of the document.

This fixes a crash when the HULLO is sent.

Use the 'standard' config.h and bootstrap the lokit.

Various things in configure.ac needed tweaking; and when at that, stop
maintaining a copy of config.h.in, and rather make sure that the
'standard' one has everything we need for the macOS app.

Copy lokit to the bundle and initialize it. Crashes early, but
something starts appearing in the webview!

Signed-off-by: Jan Holesovsky <kendy@collabora.com>
Change-Id: I6fbe7fccbc04351af216c0ff7b99f8146cafe3ef