Handle the abstract vs. regular socket names transparently
On Linux, there are 'abstract sockets' that don't have a filesystem
path. The distinction from regular unx sockets is that their name starts
with a null byte '\0', which makes the handling a bit confusing when the
code depends on a #define - the length differs, and also it is harder to
do debugging printouts.

This commit abstracts this by centralizing the adding of '\0' into one
place.

Signed-off-by: Jan Holesovsky <kendy@collabora.com>
Change-Id: I5279a85c0182620170cf0f23a1249e12837fc87c